# docs/swagger.yaml
basePath: /
definitions:
...
  api.SizeRequest:
    properties:
      count:
        type: integer
      price:
        type: number
      size:
        type: string
    type: object
...
  models.CreateMarketMessageRequest:
    properties:
      full_name:
        type: string
      message:
        type: string
      phone:
        type: string
    type: object
...
host: 192.168.55.42:8080
info:
  contact: {}
  description: API for an e-commerce platform with OTP-based authentication and product/market
    management.
  title: E-commerce API
  version: "1.0"
paths:
...
  /api/market/markets/{id}/thumbnail:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a thumbnail image for a specific market. Requires JWT authentication.
      parameters:
      - description: Market ID
        in: path
        name: id
        required: true
        type: string
      - description: Thumbnail image
        in: formData
        name: thumbnail
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Upload market thumbnail
      tags:
      - Markets
  /api/market/messages:
    post:
      consumes:
      - application/json
      description: Creates a new message from a market to superadmin. Requires market
        admin JWT authentication.
      parameters:
      - description: Message details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateMarketMessageRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new market message
      tags:
      - Market Messages
...
  /api/market/products:
    post:
      consumes:
      - multipart/form-data
      description: Creates a new product for the market admin's market with an optional
        thumbnail. Requires JWT authentication.
      parameters:
      - description: Product name
        in: formData
        name: name
        required: true
        type: string
      - description: Product name in Russian
        in: formData
        name: name_ru
        type: string
      - description: Product price
        in: formData
        name: price
        required: true
        type: number
      - description: Product discount (0-100)
        in: formData
        name: discount
        type: number
      - description: Product description
        in: formData
        name: description
        type: string
      - description: Product description in Russian
        in: formData
        name: description_ru
        type: string
      - description: Category ID
        in: formData
        name: category_id
        required: true
        type: integer
      - description: Is product active
        in: formData
        name: is_active
        type: boolean
      - description: Thumbnail image
        in: formData
        name: thumbnail
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /api/market/products/{id}/thumbnails:
    post:
      consumes:
      - multipart/form-data
      description: Adds multiple thumbnail images with associated colors to a product
        by ID. Requires JWT authentication.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Comma-separated list of colors for thumbnails
        in: formData
        name: colors
        required: true
        type: string
      - description: Comma-separated list of colors_ru for thumbnails
        in: formData
        name: colors_ru
        required: true
        type: string
      - description: Thumbnail images
        in: formData
        name: thumbnails
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Add thumbnails to a product
      tags:
      - Products
...
  /api/market/thumbnails/{thumbnail_id}/size:
    post:
      consumes:
      - application/json
      description: Adds a size with count linked to a thumbnail. Requires JWT authentication.
      parameters:
      - description: Thumbnail ID
        in: path
        name: thumbnail_id
        required: true
        type: string
      - description: Size and count
        in: body
        name: size
        required: true
        schema:
          $ref: '#/definitions/api.SizeRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Add a size by thumbnail ID
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication (e.g., "Bearer <token>")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"